stylegan2-pytorch/requirements.txt
Brotli 
certifi 
charset-normalizer 
dataclasses==0.6
filelock 
fsspec==2025.3.0
future==1.0.0
gmpy2 
idna 
Jinja2 
lmdb==1.6.2
MarkupSafe 
mkl-fft
mkl-random
mkl-service==2.4.2
mpmath 
networkx 
ninja==1.11.1.3
numpy 
pillow 
PySocks 
PyYAML
requests
six 
sympy
torch
torchaudio
torchvision
tqdm==4.67.1
triton
typing_extensions
urllib3
matplotlib
pytorch_fid 